Lab/Lab 2/lab2_1004455/requirements.txt
fastapi
redis
pillow
numpy 